+++ ui/src/views/interfaceV2/components/CreateApiModal.vue
@@ -9,49 +9,54 @@
         ref="formRef"
         :model="formState"
         :rules="rules"
-        :label-col="labelCol"
-        :wrapper-col="wrapperCol">
-      <a-form-item label="接口分类" name="tag">
-        <a-select placeholder="请选择接口分类" v-model:value="formState.tag">
-          <a-select-option value="shanghai">接口类型1</a-select-option>
-          <a-select-option value="beijing">接口类型2</a-select-option>
-        </a-select>
-      </a-form-item>
+        :label-col="{ span: 6 }"
+        :wrapper-col="{ span: 14 }">
       <a-form-item label="接口名称" name="title">
-        <a-input placeholder="接口名称" v-model:value="formState.title"/>
+        <a-input placeholder="请输入接口名称" v-model:value="formState.title"/>
       </a-form-item>
-
-      <a-form-item label="接口路径" name="path">
-        <a-input v-model:value="formState.path" class="form-item-con" placeholder="请输入接口路径">
-          <template #addonBefore>
-            <a-select
-                style="width: 120px"
-                v-model:value="formState.method"
-                :options="requestMethodOpts"
-                mode="tags"
-                placeholder="请选择请求方法">
-            </a-select>
-          </template>
-        </a-input>
+      <a-form-item label="所属分类" name="parentId">
+        <a-tree-select
+            v-model:value="formState.parentId"
+            show-search
+            :treeData="interFaceCategoryOpt"
+            style="width: 100%"
+            :dropdown-style="{ maxHeight: '400px', overflow: 'auto' }"
+            placeholder="请选择所属分类"
+            allow-clear
+            tree-default-expand-all
+        />
       </a-form-item>
-
-
+      <a-form-item label="描述" name="description">
+        <a-textarea
+            v-model:value="formState.path"
+            placeholder="清输入接口描述信息"
+            :auto-size="{ minRows: 2, maxRows: 5 }"
+        />
+      </a-form-item>
       <a-form-item :wrapper-col="{ span: 14, offset: 4 }">
         <span class="">
            注：接口请求方法可以通过详情页添加
         </span>
       </a-form-item>
-
-
     </a-form>
   </a-modal>
-
-
 </template>
 <script lang="ts" setup>
 import {ValidateErrorEntity} from 'ant-design-vue/es/form/interface';
-import {defineComponent, reactive, ref, toRaw, UnwrapRef, defineProps, defineEmits, watch} from 'vue';
-import {requestMethodOpts} from '@/config/constant';
+import {
+  reactive,
+  ref,
+  toRaw,
+  UnwrapRef,
+  defineProps,
+  defineEmits,
+  computed
+} from 'vue';
+import {useStore} from "vuex";
+import {NewInterfaceFormState} from "@/views/interfaceV2/data";
+
+const store = useStore<{ InterfaceV2, ProjectGlobal, Project }>();
+let interFaceCategoryOpt = computed<any>(() => store.state.InterfaceV2.interFaceCategoryOpt);
 
 const props = defineProps({
   visible: {
@@ -63,74 +68,40 @@ const props = defineProps({
 const emit = defineEmits(['ok', 'cancal']);
 
 
-watch(() => {
-  return props.visible
-}, () => {
-  console.log('832', props.visible)
-})
-//
-// watch(props.visible, () => {
-//   console.log('832', props.visible)
-// })
-
-/**
- *
- *   tag: '接口类型1',
- *   value: 1,
- *   title: '用户详情信息',
- *   method:'GET',
- *   path:'/api/user/:userId'
- * */
-interface FormState {
-  title: string;
-  tag: string | undefined;
-  method: string | undefined;
-  path: string | undefined;
-  value: string | undefined;
-}
-
+const formRef = ref();
 
 function ok() {
-  emit('ok',formState);
+  formRef.value
+      .validate()
+      .then(() => {
+        emit('ok', formState);
+        formRef.value.resetFields();
+      })
+      .catch((error: ValidateErrorEntity<NewInterfaceFormState>) => {
+        console.log('error', error);
+      });
 }
 
 function cancal() {
-  emit('cancal',formState);
+  emit('cancal', formState);
+  formRef.value.resetFields();
 }
 
-const formRef = ref();
-
-const formState: UnwrapRef<FormState> = reactive({
-  tag: '接口类型1',
-  value: '1',
+const formState: UnwrapRef<NewInterfaceFormState> = reactive({
   title: '',
-  method: 'GET',
-  path: ''
+  parentId: null,
+  description: '',
 });
 
 const rules = {
   title: [
     {required: true, message: '请输入接口名称', trigger: 'blur'},
-    {min: 3, max: 50, message: '最长多少个字符', trigger: 'blur'},
+    {min: 1, max: 50, message: '最少 1 个字符，最长 100 个字符', trigger: 'blur'},
   ],
-  path: [{required: true, message: 'Please select Activity zone', trigger: 'change'}],
-  tag: [{required: true, message: 'Please select activity resource', trigger: 'change'}],
+  parentId: [{required: false, message: '请选择', trigger: 'change'}],
+  description: [{required: false, message: '请输入描述', trigger: 'change'}],
 };
 
-const onSubmit = () => {
-  formRef.value
-      .validate()
-      .then(() => {
-        console.log('values', formState, toRaw(formState));
-      })
-      .catch((error: ValidateErrorEntity<FormState>) => {
-        console.log('error', error);
-      });
-};
-
-const resetForm = () => {
-  formRef.value.resetFields();
-};
 
 </script>
