+++ ui/src/views/interfaceV2/store.ts
@@ -6,23 +6,31 @@ import {InterfaceListReqParams, SaveInterfaceReqParams} from './data.d.ts';
 import {
     copyInterface,
     deleteInterface,
-    expireInterface,
+    expireInterface, getInterfaceDetail,
     getInterfaceList,
     saveInterface,
+
 } from './service';
-
-
 import {momentUtc} from "@/utils/datetime";
 import {QueryResult} from "@/views/interface/data";
+import {filterFormState} from "./data";
+import {loadCategory} from "@/services/category";
+import {getNodeMap} from "@/services/tree";
 
 export interface StateType {
     listResult: QueryResult;
+    filterState: filterFormState;
+    interFaceCategoryOpt: any,
+    interfaceDetail:any,
 }
 
 export interface ModuleType extends StoreModuleType<StateType> {
     state: StateType;
     mutations: {
         setList: Mutation<StateType>;
+        setFilterState: Mutation<StateType>;
+        setInterfaceCategory: Mutation<StateType>;
+        setInterfaceDetail: Mutation<StateType>;
     };
     actions: {
         loadList: Action<StateType, StateType>;
@@ -30,6 +38,8 @@ export interface ModuleType extends StoreModuleType<StateType> {
         disabled: Action<StateType, StateType>;
         del: Action<StateType, StateType>;
         copy: Action<StateType, StateType>;
+        loadCategory: Action<StateType, StateType>;
+        getInterfaceDetail: Action<StateType, StateType>;
     }
 }
 
@@ -44,6 +54,13 @@ const initState: StateType = {
             showQuickJumper: true,
         },
     },
+    filterState: {
+        "status": null,
+        "createUser": null,
+        "title": null,
+    },
+    interFaceCategoryOpt: [],
+    interfaceDetail:null,
 };
 
 const StoreModel: ModuleType = {
@@ -56,15 +73,28 @@ const StoreModel: ModuleType = {
         setList(state, payload) {
             state.listResult = payload;
         },
+        setFilterState(state, payload) {
+            state.filterState = payload;
+        },
+        setInterfaceCategory(state, payload) {
+            state.interFaceCategoryOpt = payload;
+        },
+        setInterfaceDetail(state, payload) {
+            state.interfaceDetail = payload;
+        },
     },
     actions: {
-        async loadList({commit, dispatch, state}, {currProjectId, page, pageSize}: any) {
+        async loadList({commit, dispatch, state}, {currProjectId, page, pageSize, opts}: any) {
             page = page || state.listResult.pagination.current;
             pageSize = pageSize || state.listResult.pagination.pageSize;
+            const otherParams = {...state.filterState, ...opts};
+
+            console.log(832, 832, opts, state.filterState)
             const res = await getInterfaceList({
                 "projectId": currProjectId,
                 "page": page,
                 "pageSize": pageSize,
+                ...otherParams,
             });
             if (res.code === 0) {
                 const {result, total} = res.data;
@@ -81,6 +111,9 @@ const StoreModel: ModuleType = {
                         "pageSize": pageSize,
                         total: total || 0,
                     },
+                });
+                commit('setFilterState', {
+                    ...otherParams
                 });
                 return true;
             } else {
@@ -117,6 +150,27 @@ const StoreModel: ModuleType = {
             const res = await copyInterface(payload.id);
             if (res.code === 0) {
                 dispatch('loadList', {currProjectId: payload.projectId});
+            } else {
+                return false
+            }
+        },
+        // 用于新建接口时选择接口分类
+        async loadCategory({commit}) {
+            const res = await loadCategory('interface');
+            if (res.code === 0) {
+                commit('setInterfaceCategory', res.data || null);
+            } else {
+                return false
+            }
+        },
+
+        // 用于新建接口时选择接口分类
+        async getInterfaceDetail({commit},payload: any) {
+            const res = await getInterfaceDetail(payload.id);
+            res.data.createdAt = momentUtc(res.data.createdAt);
+            res.data.updatedAt = momentUtc(res.data.updatedAt);
+            if (res.code === 0) {
+                commit('setInterfaceDetail', res.data || null);
             } else {
                 return false
             }
